Added CakeDC/users plugin
Installed CakeDC/Users plugin - decision was taken to use this as the
base for user related code, no point reinventing the wheel! This code
will be extended within the application to add custom functionality;
such as user registration requires admin approval
Completed the CakeDC/Users migration to add users table to the
database.